requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.3
pydantic==2.10.6
imbalanced-learn==0.12.4

httpx==0.27.2

# Optional inference acceleration runtimes/converters
onnxruntime==1.22.1
onnxmltools==1.14.0
skl2onnx==1.17.0
hummingbird-ml==0.4.12
uvloop==0.21.0
httptools==0.6.4
